+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.5.0]
-* On Integrations page, added 3rd option for authentication method, "no authentication" 
+* Re-added new Twilio regions for Virginia and Oregon Interconnect
+
+## [1.5.0]
+* On Integrations page, added 3rd option for authentication method, "no authentication"
 
 ## [1.4.9]
 * Fixed issues that were preventing Zendesk integrations from saving properly.